+++ lib/model/survey.dart
@@ -112,17 +112,9 @@ class Survey extends RuleEngine {
       : super(constants: constants, strings: strings, subRules: subRules, resultData: resultData);
 
   factory Survey.fromJson(Map<String, dynamic> json) {
-    Map<String, SurveyData> dataMap = {};
-    Map<String, dynamic> surveyDataMap = JsonUtils.mapValue(json['data']) ?? {};
-    surveyDataMap.forEach((key, value) {
-      if (value is Map<String, dynamic>) {
-        dataMap[key] = SurveyData.fromJson(key, value);
-      }
-    });
-
     return Survey(
       id: json['id'],
-      data: dataMap,
+      data: SurveyData.mapFromJson(JsonUtils.mapValue(json['data']) ?? {}),
       type: JsonUtils.stringValue(json['type']) ?? '',
       scored: JsonUtils.boolValue(json['scored']) ?? true,
       title: JsonUtils.stringValue(json['title']) ?? 'Survey',
@@ -146,7 +138,7 @@ class Survey extends RuleEngine {
     String? encodedJson = JsonUtils.encode(resultRulesJson);
     return {
       'id': id,
-      'data': JsonUtils.encodeMap(data),
+      'data': SurveyData.mapToJson(data),
       'type': type,
       'scored': scored,
       'title': title,
@@ -386,6 +378,24 @@ abstract class SurveyData {
     };
   }
 
+  static Map<String, SurveyData> mapFromJson(Map<String, dynamic> jsonMap) {
+    Map<String, SurveyData> valueMap = <String, SurveyData>{};
+    jsonMap.forEach((String key, dynamic value) {
+      if (value is Map<String, dynamic>) {
+        valueMap[key] = SurveyData.fromJson(key, value);
+      }
+    });
+    return valueMap;
+  }
+
+  static Map<String, dynamic> mapToJson(Map<String, SurveyData> valueMap) {
+    Map<String, dynamic> jsonMap = <String, dynamic>{};
+    valueMap.forEach((String key, SurveyData value) {
+      jsonMap[key] = value.toJson();
+    });
+    return valueMap;
+  }
+
   static Map<String, String> get supportedTypes => const {
     "survey_data.true_false": "True/False",
     "survey_data.multiple_choice": "Multiple Choice",
@@ -395,7 +405,7 @@ abstract class SurveyData {
     "survey_data.info": "Info",
     // "survey_data.action": "Action" // do not include because not allowed to switch to or from this type
   };
-
+  
   bool get isQuestion;
   bool get canContinue => allowSkip || response != null;
   bool get scored => scoreRule != null;
@@ -549,7 +559,7 @@ class SurveyQuestionMultipleChoice extends SurveyData {
   @override
   Map<String, dynamic> toJson() {
     Map<String, dynamic> json = baseJson();
-    json['options'] = JsonUtils.encodeList(options);
+    json['options'] = OptionData.listToJson(options);
     json['correct_answers'] = correctAnswers;
     json['allow_multiple'] = allowMultiple;
     json['self_score'] = selfScore;